% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allof_tag_api_response.R
\docType{class}
\name{AllofTagApiResponse}
\alias{AllofTagApiResponse}
\title{AllofTagApiResponse}
\format{
An \code{R6Class} generator object
}
\description{
AllofTagApiResponse Class
}
\details{
OpenAPI Petstore

This is a sample server Petstore server. For this sample, you can use the api key `special-key` to test the authorization filters.

The version of the OpenAPI document: 1.0.0
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer [optional]}

\item{\code{name}}{character [optional]}

\item{\code{code}}{integer [optional]}

\item{\code{type}}{character [optional]}

\item{\code{message}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AllofTagApiResponse-new}{\code{AllofTagApiResponse$new()}}
\item \href{#method-AllofTagApiResponse-toJSON}{\code{AllofTagApiResponse$toJSON()}}
\item \href{#method-AllofTagApiResponse-fromJSON}{\code{AllofTagApiResponse$fromJSON()}}
\item \href{#method-AllofTagApiResponse-toJSONString}{\code{AllofTagApiResponse$toJSONString()}}
\item \href{#method-AllofTagApiResponse-fromJSONString}{\code{AllofTagApiResponse$fromJSONString()}}
\item \href{#method-AllofTagApiResponse-validateJSON}{\code{AllofTagApiResponse$validateJSON()}}
\item \href{#method-AllofTagApiResponse-toString}{\code{AllofTagApiResponse$toString()}}
\item \href{#method-AllofTagApiResponse-clone}{\code{AllofTagApiResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AllofTagApiResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-AllofTagApiResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new AllofTagApiResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AllofTagApiResponse$new(
  id = NULL,
  name = NULL,
  code = NULL,
  type = NULL,
  message = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{name}}{name}

\item{\code{code}}{code}

\item{\code{type}}{type}

\item{\code{message}}{message}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AllofTagApiResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AllofTagApiResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AllofTagApiResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
AllofTagApiResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AllofTagApiResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AllofTagApiResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of AllofTagApiResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AllofTagApiResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of AllofTagApiResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AllofTagApiResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AllofTagApiResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AllofTagApiResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
AllofTagApiResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AllofTagApiResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AllofTagApiResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of AllofTagApiResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AllofTagApiResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of AllofTagApiResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AllofTagApiResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AllofTagApiResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to AllofTagApiResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AllofTagApiResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AllofTagApiResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-AllofTagApiResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AllofTagApiResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of AllofTagApiResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AllofTagApiResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AllofTagApiResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AllofTagApiResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
