% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal.R
\docType{class}
\name{Animal}
\alias{Animal}
\title{Animal}
\format{
An \code{R6Class} generator object
}
\description{
Animal Class
}
\details{
OpenAPI Petstore

This is a sample server Petstore server. For this sample, you can use the api key `special-key` to test the authorization filters.

The version of the OpenAPI document: 1.0.0
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{className}}{character}

\item{\code{color}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Animal-new}{\code{Animal$new()}}
\item \href{#method-Animal-toJSON}{\code{Animal$toJSON()}}
\item \href{#method-Animal-fromJSON}{\code{Animal$fromJSON()}}
\item \href{#method-Animal-toJSONString}{\code{Animal$toJSONString()}}
\item \href{#method-Animal-fromJSONString}{\code{Animal$fromJSONString()}}
\item \href{#method-Animal-validateJSON}{\code{Animal$validateJSON()}}
\item \href{#method-Animal-toString}{\code{Animal$toString()}}
\item \href{#method-Animal-clone}{\code{Animal$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Animal-new"></a>}}
\if{latex}{\out{\hypertarget{method-Animal-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Animal class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Animal$new(className, color = "red", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{className}}{className}

\item{\code{color}}{color. Default to 'red'.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Animal-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Animal-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Animal$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Animal in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Animal-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Animal-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Animal
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Animal$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Animal
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Animal-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Animal-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Animal$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Animal in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Animal-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Animal-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Animal
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Animal$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Animal
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Animal-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Animal-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Animal and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Animal$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Animal-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Animal-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Animal$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Animal
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Animal-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Animal-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Animal$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
