% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_of_pig.R
\docType{class}
\name{AnyOfPig}
\alias{AnyOfPig}
\title{AnyOfPig}
\format{
An \code{R6Class} generator object
}
\description{
AnyOfPig Class
}
\details{
OpenAPI Petstore

This is a sample server Petstore server. For this sample, you can use the api key `special-key` to test the authorization filters.

The version of the OpenAPI document: 1.0.0
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{any_of}}{a list of object types defined in the anyOf schema.
Initialize a new AnyOfPig.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnyOfPig-new}{\code{AnyOfPig$new()}}
\item \href{#method-AnyOfPig-fromJSON}{\code{AnyOfPig$fromJSON()}}
\item \href{#method-AnyOfPig-toJSONString}{\code{AnyOfPig$toJSONString()}}
\item \href{#method-AnyOfPig-toJSON}{\code{AnyOfPig$toJSON()}}
\item \href{#method-AnyOfPig-validateJSON}{\code{AnyOfPig$validateJSON()}}
\item \href{#method-AnyOfPig-toString}{\code{AnyOfPig$toString()}}
\item \href{#method-AnyOfPig-clone}{\code{AnyOfPig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyOfPig-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnyOfPig-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new AnyOfPig.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyOfPig$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the anyOf schemas: "BasquePig", "DanishPig"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyOfPig-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnyOfPig-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of AnyOfPig.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyOfPig$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of AnyOfPig.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyOfPig-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AnyOfPig-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize AnyOfPig to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyOfPig$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the AnyOfPig.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyOfPig-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnyOfPig-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize AnyOfPig to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyOfPig$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the AnyOfPig.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyOfPig-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnyOfPig-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to AnyOfPig and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyOfPig$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyOfPig-toString"></a>}}
\if{latex}{\out{\hypertarget{method-AnyOfPig-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyOfPig$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyOfPig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnyOfPig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyOfPig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
