% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_exception.R
\docType{class}
\name{ApiException}
\alias{ApiException}
\title{ApiException}
\format{
An \code{R6Class} generator object
}
\description{
ApiException Class
}
\details{
OpenAPI Petstore

This is a sample server Petstore server. For this sample, you can use the api key `special-key` to test the authorization filters.

The version of the OpenAPI document: 1.0.0
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{status}}{Status of the ApiException}

\item{\code{reason}}{Reason of the ApiException}

\item{\code{body}}{Body of the http response}

\item{\code{headers}}{Headers of the http response}

\item{\code{errorObject}}{error object type}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ApiException-new}{\code{ApiException$new()}}
\item \href{#method-ApiException-toString}{\code{ApiException$toString()}}
\item \href{#method-ApiException-clone}{\code{ApiException$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiException-new"></a>}}
\if{latex}{\out{\hypertarget{method-ApiException-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ApiExceptino class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiException$new(status = NULL, reason = NULL, http_response = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{status}}{HTTP status.}

\item{\code{reason}}{Reason of the ApiException.}

\item{\code{http_response}}{HTTP response object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiException-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ApiException-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string format of ApiException.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiException$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the string format of ApiException.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiException-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ApiException-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiException$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
