% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_all_of.R
\docType{class}
\name{CatAllOf}
\alias{CatAllOf}
\title{CatAllOf}
\format{
An \code{R6Class} generator object
}
\description{
CatAllOf Class
}
\details{
OpenAPI Petstore

This is a sample server Petstore server. For this sample, you can use the api key `special-key` to test the authorization filters.

The version of the OpenAPI document: 1.0.0
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{declawed}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CatAllOf-new}{\code{CatAllOf$new()}}
\item \href{#method-CatAllOf-toJSON}{\code{CatAllOf$toJSON()}}
\item \href{#method-CatAllOf-fromJSON}{\code{CatAllOf$fromJSON()}}
\item \href{#method-CatAllOf-toJSONString}{\code{CatAllOf$toJSONString()}}
\item \href{#method-CatAllOf-fromJSONString}{\code{CatAllOf$fromJSONString()}}
\item \href{#method-CatAllOf-validateJSON}{\code{CatAllOf$validateJSON()}}
\item \href{#method-CatAllOf-toString}{\code{CatAllOf$toString()}}
\item \href{#method-CatAllOf-clone}{\code{CatAllOf$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CatAllOf-new"></a>}}
\if{latex}{\out{\hypertarget{method-CatAllOf-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new CatAllOf class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatAllOf$new(declawed = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{declawed}}{declawed}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CatAllOf-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CatAllOf-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatAllOf$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CatAllOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CatAllOf-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CatAllOf-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of CatAllOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatAllOf$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of CatAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CatAllOf-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CatAllOf-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatAllOf$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CatAllOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CatAllOf-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CatAllOf-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of CatAllOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatAllOf$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of CatAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CatAllOf-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CatAllOf-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to CatAllOf and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatAllOf$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CatAllOf-toString"></a>}}
\if{latex}{\out{\hypertarget{method-CatAllOf-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatAllOf$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of CatAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CatAllOf-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CatAllOf-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CatAllOf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
