% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/danish_pig.R
\docType{class}
\name{DanishPig}
\alias{DanishPig}
\title{DanishPig}
\format{
An \code{R6Class} generator object
}
\description{
DanishPig Class
}
\details{
OpenAPI Petstore

This is a sample server Petstore server. For this sample, you can use the api key `special-key` to test the authorization filters.

The version of the OpenAPI document: 1.0.0
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{className}}{character}

\item{\code{size}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DanishPig-new}{\code{DanishPig$new()}}
\item \href{#method-DanishPig-toJSON}{\code{DanishPig$toJSON()}}
\item \href{#method-DanishPig-fromJSON}{\code{DanishPig$fromJSON()}}
\item \href{#method-DanishPig-toJSONString}{\code{DanishPig$toJSONString()}}
\item \href{#method-DanishPig-fromJSONString}{\code{DanishPig$fromJSONString()}}
\item \href{#method-DanishPig-validateJSON}{\code{DanishPig$validateJSON()}}
\item \href{#method-DanishPig-toString}{\code{DanishPig$toString()}}
\item \href{#method-DanishPig-clone}{\code{DanishPig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DanishPig-new"></a>}}
\if{latex}{\out{\hypertarget{method-DanishPig-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DanishPig class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DanishPig$new(className, size, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{className}}{className}

\item{\code{size}}{size}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DanishPig-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DanishPig-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DanishPig$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
DanishPig in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DanishPig-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DanishPig-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of DanishPig
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DanishPig$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of DanishPig
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DanishPig-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DanishPig-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DanishPig$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
DanishPig in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DanishPig-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DanishPig-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of DanishPig
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DanishPig$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of DanishPig
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DanishPig-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DanishPig-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to DanishPig and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DanishPig$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DanishPig-toString"></a>}}
\if{latex}{\out{\hypertarget{method-DanishPig-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DanishPig$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of DanishPig
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DanishPig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DanishPig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DanishPig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
