% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_all_of.R
\docType{class}
\name{DogAllOf}
\alias{DogAllOf}
\title{DogAllOf}
\format{
An \code{R6Class} generator object
}
\description{
DogAllOf Class
}
\details{
OpenAPI Petstore

This is a sample server Petstore server. For this sample, you can use the api key `special-key` to test the authorization filters.

The version of the OpenAPI document: 1.0.0
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{breed}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DogAllOf-new}{\code{DogAllOf$new()}}
\item \href{#method-DogAllOf-toJSON}{\code{DogAllOf$toJSON()}}
\item \href{#method-DogAllOf-fromJSON}{\code{DogAllOf$fromJSON()}}
\item \href{#method-DogAllOf-toJSONString}{\code{DogAllOf$toJSONString()}}
\item \href{#method-DogAllOf-fromJSONString}{\code{DogAllOf$fromJSONString()}}
\item \href{#method-DogAllOf-validateJSON}{\code{DogAllOf$validateJSON()}}
\item \href{#method-DogAllOf-toString}{\code{DogAllOf$toString()}}
\item \href{#method-DogAllOf-clone}{\code{DogAllOf$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DogAllOf-new"></a>}}
\if{latex}{\out{\hypertarget{method-DogAllOf-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DogAllOf class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DogAllOf$new(breed = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{breed}}{breed}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DogAllOf-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DogAllOf-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DogAllOf$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
DogAllOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DogAllOf-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DogAllOf-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of DogAllOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DogAllOf$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of DogAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DogAllOf-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DogAllOf-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DogAllOf$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
DogAllOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DogAllOf-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DogAllOf-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of DogAllOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DogAllOf$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of DogAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DogAllOf-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DogAllOf-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to DogAllOf and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DogAllOf$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DogAllOf-toString"></a>}}
\if{latex}{\out{\hypertarget{method-DogAllOf-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DogAllOf$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of DogAllOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DogAllOf-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DogAllOf-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DogAllOf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
