% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_one_of.R
\docType{class}
\name{NestedOneOf}
\alias{NestedOneOf}
\title{NestedOneOf}
\format{
An \code{R6Class} generator object
}
\description{
NestedOneOf Class
}
\details{
OpenAPI Petstore

This is a sample server Petstore server. For this sample, you can use the api key `special-key` to test the authorization filters.

The version of the OpenAPI document: 1.0.0
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{size}}{integer [optional]}

\item{\code{nested_pig}}{\link{Pig} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NestedOneOf-new}{\code{NestedOneOf$new()}}
\item \href{#method-NestedOneOf-toJSON}{\code{NestedOneOf$toJSON()}}
\item \href{#method-NestedOneOf-fromJSON}{\code{NestedOneOf$fromJSON()}}
\item \href{#method-NestedOneOf-toJSONString}{\code{NestedOneOf$toJSONString()}}
\item \href{#method-NestedOneOf-fromJSONString}{\code{NestedOneOf$fromJSONString()}}
\item \href{#method-NestedOneOf-validateJSON}{\code{NestedOneOf$validateJSON()}}
\item \href{#method-NestedOneOf-toString}{\code{NestedOneOf$toString()}}
\item \href{#method-NestedOneOf-clone}{\code{NestedOneOf$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NestedOneOf-new"></a>}}
\if{latex}{\out{\hypertarget{method-NestedOneOf-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new NestedOneOf class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NestedOneOf$new(size = NULL, nested_pig = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{size}

\item{\code{nested_pig}}{nested_pig}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NestedOneOf-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NestedOneOf-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NestedOneOf$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NestedOneOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NestedOneOf-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NestedOneOf-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of NestedOneOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NestedOneOf$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of NestedOneOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NestedOneOf-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-NestedOneOf-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NestedOneOf$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NestedOneOf in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NestedOneOf-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-NestedOneOf-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of NestedOneOf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NestedOneOf$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of NestedOneOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NestedOneOf-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NestedOneOf-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to NestedOneOf and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NestedOneOf$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NestedOneOf-toString"></a>}}
\if{latex}{\out{\hypertarget{method-NestedOneOf-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NestedOneOf$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of NestedOneOf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NestedOneOf-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NestedOneOf-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NestedOneOf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
