% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\docType{class}
\name{Order}
\alias{Order}
\title{Order}
\format{
An \code{R6Class} generator object
}
\description{
Order Class
}
\details{
OpenAPI Petstore

This is a sample server Petstore server. For this sample, you can use the api key `special-key` to test the authorization filters.

The version of the OpenAPI document: 1.0.0
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer [optional]}

\item{\code{petId}}{integer [optional]}

\item{\code{quantity}}{integer [optional]}

\item{\code{shipDate}}{character [optional]}

\item{\code{status}}{character [optional]}

\item{\code{complete}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Order-new}{\code{Order$new()}}
\item \href{#method-Order-toJSON}{\code{Order$toJSON()}}
\item \href{#method-Order-fromJSON}{\code{Order$fromJSON()}}
\item \href{#method-Order-toJSONString}{\code{Order$toJSONString()}}
\item \href{#method-Order-fromJSONString}{\code{Order$fromJSONString()}}
\item \href{#method-Order-validateJSON}{\code{Order$validateJSON()}}
\item \href{#method-Order-toString}{\code{Order$toString()}}
\item \href{#method-Order-clone}{\code{Order$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Order-new"></a>}}
\if{latex}{\out{\hypertarget{method-Order-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Order class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Order$new(
  id = NULL,
  petId = NULL,
  quantity = NULL,
  shipDate = NULL,
  status = NULL,
  complete = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{petId}}{petId}

\item{\code{quantity}}{quantity}

\item{\code{shipDate}}{shipDate}

\item{\code{status}}{Order Status}

\item{\code{complete}}{complete. Default to FALSE.}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Order-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Order-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Order$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Order in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Order-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Order-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Order
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Order$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Order
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Order-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Order-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Order$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Order in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Order-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Order-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Order
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Order$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Order
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Order-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Order-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Order and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Order$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Order-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Order-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Order$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Order
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Order-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Order-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Order$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
