% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pig.R
\docType{class}
\name{Pig}
\alias{Pig}
\title{Pig}
\format{
An \code{R6Class} generator object
}
\description{
Pig Class
}
\details{
OpenAPI Petstore

This is a sample server Petstore server. For this sample, you can use the api key `special-key` to test the authorization filters.

The version of the OpenAPI document: 1.0.0
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of object types defined in the oneOf schema.
Initialize a new Pig.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Pig-new}{\code{Pig$new()}}
\item \href{#method-Pig-fromJSON}{\code{Pig$fromJSON()}}
\item \href{#method-Pig-toJSONString}{\code{Pig$toJSONString()}}
\item \href{#method-Pig-toJSON}{\code{Pig$toJSON()}}
\item \href{#method-Pig-validateJSON}{\code{Pig$validateJSON()}}
\item \href{#method-Pig-toString}{\code{Pig$toString()}}
\item \href{#method-Pig-clone}{\code{Pig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pig-new"></a>}}
\if{latex}{\out{\hypertarget{method-Pig-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Pig.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pig$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "BasquePig", "DanishPig"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pig-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Pig-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Pig.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pig$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of Pig.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pig-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Pig-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize Pig to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pig$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the Pig.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pig-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Pig-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize Pig to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pig$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the Pig.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pig-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Pig-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to Pig and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pig$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pig-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Pig-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pig$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Pig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
