% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\docType{class}
\name{Special}
\alias{Special}
\title{Special}
\format{
An \code{R6Class} generator object
}
\description{
Special Class
}
\details{
OpenAPI Petstore

This is a sample server Petstore server. For this sample, you can use the api key `special-key` to test the authorization filters.

The version of the OpenAPI document: 1.0.0
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{item_self}}{integer [optional]}

\item{\code{item_private}}{character [optional]}

\item{\code{item_super}}{character [optional]}

\item{\code{123_number}}{character [optional]}

\item{\code{array[test]}}{character [optional]}

\item{\code{empty_string}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Special-new}{\code{Special$new()}}
\item \href{#method-Special-toJSON}{\code{Special$toJSON()}}
\item \href{#method-Special-fromJSON}{\code{Special$fromJSON()}}
\item \href{#method-Special-toJSONString}{\code{Special$toJSONString()}}
\item \href{#method-Special-fromJSONString}{\code{Special$fromJSONString()}}
\item \href{#method-Special-validateJSON}{\code{Special$validateJSON()}}
\item \href{#method-Special-toString}{\code{Special$toString()}}
\item \href{#method-Special-clone}{\code{Special$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Special-new"></a>}}
\if{latex}{\out{\hypertarget{method-Special-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Special class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Special$new(
  item_self = NULL,
  item_private = NULL,
  item_super = NULL,
  `123_number` = NULL,
  `array[test]` = NULL,
  empty_string = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{item_self}}{item_self}

\item{\code{item_private}}{item_private}

\item{\code{item_super}}{item_super}

\item{\code{123_number}}{123_number}

\item{\code{array[test]}}{array[test]}

\item{\code{empty_string}}{empty_string}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Special-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Special-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Special$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Special in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Special-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Special-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Special
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Special$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Special
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Special-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Special-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Special$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Special in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Special-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Special-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Special
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Special$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Special
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Special-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Special-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Special and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Special$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Special-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Special-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Special$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Special
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Special-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Special-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Special$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
