% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_pet_request.R
\docType{class}
\name{UpdatePetRequest}
\alias{UpdatePetRequest}
\title{UpdatePetRequest}
\format{
An \code{R6Class} generator object
}
\description{
UpdatePetRequest Class
}
\details{
OpenAPI Petstore

This is a sample server Petstore server. For this sample, you can use the api key `special-key` to test the authorization filters.

The version of the OpenAPI document: 1.0.0
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{jsonData}}{\link{Pet} [optional]}

\item{\code{binaryDataN2Information}}{data.frame [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UpdatePetRequest-new}{\code{UpdatePetRequest$new()}}
\item \href{#method-UpdatePetRequest-toJSON}{\code{UpdatePetRequest$toJSON()}}
\item \href{#method-UpdatePetRequest-fromJSON}{\code{UpdatePetRequest$fromJSON()}}
\item \href{#method-UpdatePetRequest-toJSONString}{\code{UpdatePetRequest$toJSONString()}}
\item \href{#method-UpdatePetRequest-fromJSONString}{\code{UpdatePetRequest$fromJSONString()}}
\item \href{#method-UpdatePetRequest-validateJSON}{\code{UpdatePetRequest$validateJSON()}}
\item \href{#method-UpdatePetRequest-toString}{\code{UpdatePetRequest$toString()}}
\item \href{#method-UpdatePetRequest-clone}{\code{UpdatePetRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdatePetRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-UpdatePetRequest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new UpdatePetRequest class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdatePetRequest$new(jsonData = NULL, binaryDataN2Information = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{jsonData}}{jsonData}

\item{\code{binaryDataN2Information}}{binaryDataN2Information}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdatePetRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UpdatePetRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdatePetRequest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UpdatePetRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdatePetRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UpdatePetRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of UpdatePetRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdatePetRequest$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UpdatePetRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdatePetRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UpdatePetRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdatePetRequest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UpdatePetRequest in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdatePetRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UpdatePetRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of UpdatePetRequest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdatePetRequest$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of UpdatePetRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdatePetRequest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UpdatePetRequest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to UpdatePetRequest and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdatePetRequest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdatePetRequest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-UpdatePetRequest-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdatePetRequest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of UpdatePetRequest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UpdatePetRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UpdatePetRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UpdatePetRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
