% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\docType{class}
\name{User}
\alias{User}
\title{User}
\format{
An \code{R6Class} generator object
}
\description{
User Class
}
\details{
OpenAPI Petstore

This is a sample server Petstore server. For this sample, you can use the api key `special-key` to test the authorization filters.

The version of the OpenAPI document: 1.0.0
Generated by: https://openapi-generator.tech
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer [optional]}

\item{\code{username}}{character [optional]}

\item{\code{firstName}}{character [optional]}

\item{\code{lastName}}{character [optional]}

\item{\code{email}}{character [optional]}

\item{\code{password}}{character [optional]}

\item{\code{phone}}{character [optional]}

\item{\code{userStatus}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-User-new}{\code{User$new()}}
\item \href{#method-User-toJSON}{\code{User$toJSON()}}
\item \href{#method-User-fromJSON}{\code{User$fromJSON()}}
\item \href{#method-User-toJSONString}{\code{User$toJSONString()}}
\item \href{#method-User-fromJSONString}{\code{User$fromJSONString()}}
\item \href{#method-User-validateJSON}{\code{User$validateJSON()}}
\item \href{#method-User-toString}{\code{User$toString()}}
\item \href{#method-User-clone}{\code{User$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-User-new"></a>}}
\if{latex}{\out{\hypertarget{method-User-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new User class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$new(
  id = NULL,
  username = NULL,
  firstName = NULL,
  lastName = NULL,
  email = NULL,
  password = NULL,
  phone = NULL,
  userStatus = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{username}}{username}

\item{\code{firstName}}{firstName}

\item{\code{lastName}}{lastName}

\item{\code{email}}{email}

\item{\code{password}}{password}

\item{\code{phone}}{phone}

\item{\code{userStatus}}{User Status}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-User-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-User-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
User in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-User-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-User-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of User
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of User
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-User-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-User-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
User in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-User-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-User-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of User
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of User
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-User-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-User-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to User and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-User-toString"></a>}}
\if{latex}{\out{\hypertarget{method-User-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of User
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-User-clone"></a>}}
\if{latex}{\out{\hypertarget{method-User-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{User$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
